+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the gtk-fortran project are documented in this file. The 
 
 ### Added
 - New types (enums) in `src/cfwrapper/scan_types_and_enums.py`, useful for other GNOME libraries.
+- Initializes `types_enums.gtk_funptr` list in `src/cfwrapper/scan_types_and_enums.py` with a few funptr types, and removes possibly duplicated ones.
 
 
 ## [gtk-fortran 4.4.0] 2023-05-02